Add security policy document
While it is unlikely that mantid will have a direct security issue, it
is useful to publicly display a policy for how to report issues and how
they will be handled. This wording borrows ideas from
[prefix-dev/rattler-build](https://github.com/prefix-dev/rattler-build/blob/main/SECURITY.md) and [microsoft/repo-templates](https://github.com/microsoft/repo-templates/blob/main/shared/SECURITY.md).

This is suggested in the [openssf best practices
badge](https://www.bestpractices.dev/en/projects/4316).