Check for none when building child algorithms from dicts in python (#39820)
### Description of work
When trying to execute a child algorithm from a python algorithm,
without using the simpleapi, it is possible to pass a dictionary of
kwargs to the method 'createChildAlgorithm' , which is interfaced from
the algorithm class.
The translation from the python properties to the cpp property types is
done in the algorithm exports definition file. Before assigning the
property, there is a check with the not(!) cpp operator, I guess to
assess whether there is some error on exporting the python object or if
the property has been declared as none, and continues iterating over
other properties if that's the case. But boost::python api checks for
truthiness of the contained boost::python object and as a result some
perfectly valid properties (like `WorkspaceIndex`=0) are not set in the
algorithm, as they are evaluated to be false before assigning.
An example may illustrate better: 
```
from mantid.simpleapi import *
from mantid.api import AlgorithmFactory, PythonAlgorithm, AlgorithmManager

class TestChild(PythonAlgorithm):

    def category(self):
        return "Utility"

    def name(self):
        return "TestChild"
    
    def PyInit(self):
        pass

    def PyExec(self):
        alg =self.createChildAlgorithm("CreateSampleWorkspace", XMax=0.0)
        alg.initialize()
        print("Is it 0? ->",alg.getPropertyValue("XMax"))

AlgorithmFactory.subscribe(TestChild)

asChildAlg = AlgorithmManager.createUnmanaged("TestChild")
asChildAlg.initialize()
asChildAlg.execute()

testAlg = AlgorithmManager.createUnmanaged("CreateSampleWorkspace")
testAlg.initialize()
testAlg.setProperty("XMax", 0.0)
print("Is it 0? ->",testAlg.getPropertyValue("XMax"))
```
A test python algorithm is created and within it, a child algorithm for
`CreateSampleWorkspace` is called with the property `XMax` set to 0. The
same is done calling directly the algorithm. In the first case, the
property is not set, so we retrieve the default value for this property
(2000), while in the second case we get 0.

To fix this, I have changed to just check whether the property is equal
to none, just being careful that `StoreInADS` is handled before reaching
this point (which shouldn't be a concern when calling a child algorithm
from simpleapi anyway)



<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

Build the branch and check the example of the pr, now the result should
be similar in both the child algorithm and the main algorithm. (0)

_No release notes as it is an internal change, very unlikely to have
impacted the user_