Add partial functionality to visualisation features in plotting tab (#39612)
This PR contains changes that activate functionality of several buttons
in Plotting tab of Single Crystal Elastic mode of the DNS Reduction
interface. In particular, the following buttons are now functional:
1) "Grid selector", 
2) "Grid style selector", 
3) "Font size selector", 
4) "Reverse color map".

This PR includes a second part of adding advanced plotting capabilities
for the Single Cristal Elastic mode of DNS Reduction GUI. Complete
integration of the Single Cristal Elastic mode into the DNS Reduction
GUI will be implemented in several steps (in order to ensure reasonable
PR sizes and reduce the workload for the reviewers).

Fixes part 6b of #36407