Use a stamp file to prevent multiple installs of the same pip package (#39878)
### Description of work

This prevents our build process continually installing/uninstalling
editable python packages.

The custom command we add to create packages should not run if the
packages already exist. This is however not working/is brittle. This is
currently leading now failures in the windows build and test stage of
the nightly pipeline due to ninja parallelisation.

This PR changes the output of the cmake command to a simple stamp file,
with the previous outputs listed as byproducts. If the stamp file
exists, the install function will not run again.

### To test:
- CI tests pass
- Linux build and test passes:
https://builds.mantidproject.org/job/build_packages_from_branch/1335/

I intended to run a windows build and test but didn't. I propose we
merge this regardless, as the worst that can occur is that we break the
already broken windows build. Windows b&t:
https://builds.mantidproject.org/job/build_packages_from_branch/1339/

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->