Allow a Separate High Angle Bank Scale Factor to Be Defined in User File (#39806)
### Description of work

Adds the option to supply a `front_scale` value in the user file. This
overrides a value given to `rear_scale` to be applied to the HAB
workspaces.

Note: This PR modifies SANS TOML V1. Given that this parameter was
already mentioned in the documentation for TOML V1 but didn't do
anything, the decision was made to update the old format rather than
updating

Also fixes a test class for the beam centre finder that wasn't in the
CMakeLists.txt file and so wasn't being run.

Closes #39341 
Closes #27948 

### To test:
1. Make a copy of the `MaskFile.toml` User file from the `loqdemo`.
3. In this copy, add a `front_scale` value to the
`[detector.configuration]` section.
4. Launch workbench and open the ISIS SANS GUI
5. Load the original `MaskFile.toml` and `batch_mode_reduction.csv`
files.
6. On the `second_time` run, enter the path to the `front_scale` copy
into the `User File` box in the grid.
7. Run Everything
8. Plot the output files for the HAB and LAB onto separate plots for one
of the runs
9. Drag over the workspaces for the other run to overplot them
10. Check that the LAB curves are the same and the HAB curve is using
the scale you provided in the user file.