Abins systemtests cleanup (#39739)
Tidy up and simplify Abins system tests. With a better base class setup
we need a lot less boilerplate and the tests themselves can call Abins
algorithm directly without too many lines of code. This seems a better
verbosity/clarity tradeoff than the existing design.

This should make the system tests faster and easier to debug, fix and
extend.

Some slow system tests were disabled for pull-requests, only
running during full checks. These run quite quickly now due to
performance improvements in Abins, so have been re-enabled.

#### Further detail of work
The existing implementation of Abins systemtests involves a lot of
boilerplate. The calls to Abins algorithms are abstracted by method
calls, and parameters are set individually following a kind of "builder"
pattern.

That pattern provides some validation of appropriate parameters, but a
test suite is not a good place to have such validation. We are supposed
to be testing the behaviour of the production code when it is fed
whatever paramteters we choose to test!

By providing a default dictionary attribute and using compact `{some:
dict} | {data: update}` notation to produce modified copies, we can make
direct calls to Abins that express their intent clearly and avoid the
need to read back-and-forth in the file to make sense of each test.