AlignAndFocusPowderSlim: Only load the 6 banks since everything else assumes we have 6 (#39906)
### Description of work

The prototype algorithm
[AlignAndFocusPowderSlim](https://docs.mantidproject.org/nightly/algorithms/AlignAndFocusPowderSlim-v1.html)
assume we are only loading 6 banks. So for now don't try loading any
other banks.

As started in the docs "hard coded for 6 particular groups". This is to
address the issue of new files having additional banks.

This algorithm is under heavy development and we need it to just work
for testing.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

This should now be able to load newer files like `VULCAN_260351`.

*This does not require release notes* because this is a prototype
algorithm that is now doing what the docs say.

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->