Dgsplanner updates (#39905)
The following issues are addressed in this PR:

1. If crystallographic convention is used, it prints it as a title to the plot
2. Masking an empty instrument with a workspace2d from real data (without monitors) causes MaskDetectors to be slow

### To test:

Start Mantid. Use a processed vanadium file (say from SEQUOIA) as a mask
file. Just click plot. The main branch is extremely slow, this one is
less than 1 second. Change the Q convention to "Crytallography" in
settings. A title should show up in the plot, warning about the change