Performance improvement for poldi autocorrelation and simulation (#39776)
Performance improvements required for when POLDI gets an additional
detector bank and starts measuring (both banks) with x10 more detector
channels.

Joblib multithreading has been used in `PoldiAutoCorrelation` and the
code to simulate the 2D POLDI workspace.
A further speed up in `PoldiAutoCorrelation` is provided by adding an
option to use nearest interpolation - the existing behaviour (linear
interpolation) in the default. Smaller improvements have also been made
by e.g. using `np.take` with `mode='wrap'` to avoid taking the modulus
with cycle time etc.

The nearest interpolation provides a speed up of ~2-3, in principle it
could be less accurate but in practice it seems to average out over all
the spectra and tyhe peak positions agree for both interpolation
methods. The only noticeable difference is that the low-Q part of the
spectrum in slightly noisier.