Catch runtime errors when adding period logs (#39870)
Inconsistencies sometimes occur in period-related logs within Event
NeXus files, such as duplicated entries with the same timestamps.
Therefore, these repeated entries are reduced to a single entry. This
can lead to runtime errors when loading the file because it is not
possible to determine the ending value from a TimeSeriesProperty with a
single time entry. In the past, this has resulted in Mantid crashes. In
order to prevent this, the related event workspace is now created
without the period logs if issues are detected.