AlignAndFocusPowderSlim selective sample log loading (#39934)
To follow behavior from
[AlignAndFocusPowderFromFiles](https://docs.mantidproject.org/nightly/algorithms/AlignAndFocusPowderFromFiles-v1.html),
this adds the mutually exclusive ``LogAllowList`` and ``LogBlockList``
parameters. Fortunately, the underlying code for loading logs already
had an entrypoint for the feature.

Separately, the unit test was refactored to make the settings for an
individual run of the algorithm much clearer. The `TestConfig` object
was introduced rather than an increasing list of options to a function.

There is no associated issue, but this is for
[EWM12971](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20(Change%20Management)#action=com.ibm.team.workitem.viewWorkItem&id=12971).

### To test:

Use an existing VULCAN file and supply the `LogBlockList` property to
skip `*skf*` logs. In the "view sample logs" those logs should be
missing.