Add new output options to PlotPeakByLogValue algorithm and Sequetial Fit dialouge (#39605)
### Description of work
This PR add two new options to the PlotPeakByLogValue algorithm that are
enabled when the "CreateOutput" option is true. First option
"AppendIdxToOutputName", appends the index of the corresponding output
depending on which format is used in the input (either spectrum number,
workspace index, or numeric value). This is enabled now by default for
the sequential fit output. The second option, "Output2D", creates a 2D
workspace of the results table, displaying the workspace index versus
parameter value. This option groups the parameter results together to
create a workspace index vs parameters
It also adds a new UI option to the Sequential fit dialogue to enable
this new 2D option.

Fixes #39294

<!-- If the original issue was raised by a user they should be named
here. Do not leak email addresses
**Report to:** [user name]
-->

#### Further detail of work
<!-- Please provide a more detailed description of the work that has
been undertaken.
-->
The implementation involves enabling the `alwaysStoreInAds` option in
the child Fit algorithms. It turned out that because this option was not
enabled, the output workspace returned by the Fit algorithm was an empty
string. This is why the output names are the same as the group workspace
name, numbered by their position. One thing to note is that the final
outputs seem to be duplicated using both the workspace names and
workspace objects. This is because if you enable `alwaysStoreInAds`,
then the returned workspace objects from the algorithm will be null, and
the only way to access the workspaces is from the ADS. In the case of
`alwaysStoreInAds` being false, the returned workspaces will contain the
actual workspace, but their names will be null. This is one of the
drawbacks of using ADS vs not.

### To test:
1- Load any data (e.g. iris26176_graphite002_red)
2- Go to Plot Spectrum and click Fit 
3- Add a peak (e.g. Lorentzian)
4- Press Sequential Fit, and a dialogue will open
5- Check the `Create Output` check box
6- In the table view, set `Range` to any preferred workspace index range
(e.g. 4-6)
7- Press Fit
8- You will see 3 output group workspaces suffixed with "_Parameters",
"_NormalisedCovarianceMatrices" and "_Workspaces"
9- Each group workspace will have the same number of spectra in the
selected range
10- You can see that each workspace name includes the spectrum number
and is suffixed with "_Parameters", "_NormalisedCovarianceMatrix" and
"_Workspace" in each group, respectively (e.g.
iris26176_graphite002_red_res_3_Parameters)

11- Try to enable the new "Output spectrum vs params" in the sequential
fit dialogue, and a new output suffixed with "_ws" will appear. Plot the
workspace, and it will contain a plot for each parameter.