Fix Docs Packaging (`mambabuild`) Job Failing to Find & Generate Test Data (#39924)
The packaging steps are run using `mambabuild`, which uses the scripts
in `conda/recipes/<PACKAGE>/build.sh`. This is different to the normal
PR build and test runs which use the `build` script
(`buildconfig/Jenkins/Conda/build`).

When running with `mambabuild`, environment variables (such as
`MANTID_DATA_STORE`) that are set in the Node Configuration on Jenkins
are ignored. Thus the data store defaults to being in
`/home/user/MantidExternalData/` rather than the expected
`/mantid_data/`.

Some recent-ish change (either in docker or Jenkins) has altered the
permissions of the Jenkins agent process, such that it can no longer
edit/access the contents of the external data cache in the home
directory.

This PR explicitly adds this environment variable for Linux so that the
packaging PR and Nightly jobs use the external data volume defined in
the Jenkins node configuration.

Closes #39916