Handle instruments with only a subset of detectors
At least one instrument, WISH, has multiple detector IDs at the same
position. We only plot one of those, currently the first, but the
side-by-side projection needs to get the relevant component indices for
the given detectors. This is the index of the given detectors in the
list of all the detectors.