new bilby loader (#39928)
The ANSTO Bilby instrument has had an upgrade and is now able to save a
new file format, in parallel to the existing file, that supports
timestamping of all events. This submission is a loader for the new file
format so that scientists can transition to the new file structure and
it assess it's new capabilities. The submission includes minimal unit
and doc tests to demonstrate that it is ready to be evaluated.