Acccount for ragged workspaces when rebinning
Calling extractX on a ragged workspace will truncate the bin edges to
the number of bin edges of the spectra with the fewest edges. So we
need to loop over each spectra individually to get the rebin parameters.

Also, no need to convert units if the workspace is already in d-Spacing.