Fix Help Window path (#39826)
### Description of work
This PR fixes the path of the docs that caused an issue with the offline
and online modes in the new Python help window. It also fixes a problem
with Qt icons not being shown correctly on Windows. This issue doesn't
resolve the conda installation bug on Windows.

Closes #39506. <!-- One line per closed issue. -->

### To test:
- Development Build
1- Build this branch locally
2- Go to Mantid Help 
3- You will see the help docs in online mode and a warning that the docs
were not found in the local path
4- Build docs-html
5- Rerun Mantid; it should now show the offline docs 
6- Check that all icons are shown correctly

- Release Build
Check the unstable build is working as expected
https://builds.mantidproject.org/job/build_packages_from_branch/1299

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->