Add interactor styles to enable rectangle picking in 2d
These interactor styles allow for rectangular picking with right button
on both 3D and 2D projections.

On the 2D projection case, this is achieved by layering two vtk styles
on top of the interactor, and disabling certain observers from both
styles by overwriting them to do nothing.

As a word of caution, calling `removeObserver` or `removeAllObservers`
on a vtk style does not remove default observers, the only way that I
found around this is to instead overwrite default observers using
`AddObserver`.