Add miniplot toolbar and correct units (#39978)
Does two things in the new Instrument View:

- Adds a toolbar to the 'miniplot' 
- Displays the correct units for the integration slider

Closes #39748. <!-- One line per closed issue. -->

Several functions need to be renamed because they call say 'TOF' when
the unit of the workspace could be anything, but this can be done in
maintenance in a separate PR.

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

Check correct unit label is displayed by opening workspaces in different
units in the new Instrument View.

Test out the new buttons on the line plot showing the selected spectra
(or their sum).

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->