Fix for embedded plot warning in inelastic/indirect interfaces (#39979)
This bug is quite sneaky. The problem seems to appear when clicking down
horizontal or vertical markers of embedded plots (for example the ones
in the symmetrise interface). This in fact tries to determine if the
mouse position is above the marker by making a data transform using
matplotlib. The returned transformed position and the marker position
are compared within a sensitivity margin. This has a np.bool return
type. Now, I don't know how the boost python api takes this np.bool and
produces the warning exactly, but that seems to be the problem. I have
modified the return type of the function to be just a python native
boolean type. In any case is probably better to not have mixed return
types in the boost python call from cpp. The warning does not appear
anymore.