Add section for missing release notes. (#39981)
### Description of work
A new directory was added at some point for some release notes. The
directory wasn't included in the framework index. This adds it so that
the release notes are not lost.

### To test:
Build docs and check that the release notes are present. I've done it
for you so you don't have to:
<img width="524" height="394" alt="image"
src="https://github.com/user-attachments/assets/a58a8846-b51c-458d-971e-22f91fdc34ca"
/>