Delete FindPoco.cmake
Poco comes with its own cmake module, we can use that instead. The only
thing is that it defines UNICODE and _UNICODE as compiler arguments,
which causes chaos on Windows, since we would have to use wstring in
several places. Hence in the cmake the UNICODE and _UNICODE flags are
removed.