Refactor PeakShapeEllipsoid from std::vector to std::array
Since the underlying objects for Ellipsoids are always length 3,
change the code to use fixed length arrays. There are various knock-on
effects for other code that intersects this, but it should allow the
compiler to deal with memory more efficiently which should have a positive
effect on performance with a larger number of peaks (think protein
crystallography).