Create new ProcessBankSplitTask to handle loading data and splitting to correct output workspace.
Instead of loading data for each output workspace in turn this now just loads all the data and split the data to the correct workspace using ranges library. This doesn't cause the data to be copied as it's just a view on the data.

This required changing ProcessEventsTask to be templated so it could also accept the std::ranges::views. The implementation has to be moved to the header so it would work with these complex types.