Fix for show invisible workspace checkbox unreliable behaviour (#40036)
### Description of work
This PR fixes a bug in setting the ``Show invisible Workspaces``
checkbox in the settings menu. The checkbox behaviour is unreliable, and
under some conditions: #40028 , it fails to properly update the setting.
I think the problem was in the stage of loading the interface, as the
initial value of the setting was compared against the string `true`
while in most of the other parts of mantid (including the same presenter
for the settings interface) the invisible workspace setting is compared
against '1' or '0'. I have leave it as '1' or '0', although this could
be converted to be 'true' or 'false', as it does not matter when setting
the property from the config service.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes #40028. <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

- The bug reported in #40028 does not reproduce after this fix. 

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->