Remove osx 64 from main nightly (#40077)
### Description of work
An attempt was made to remove osx-64 from the main nightly but it was
broken. I have reverted those changes and implemented the correct
removal. Since this is going into `main` it will not impact the
`release-next` builds until we enter code freze for v6.15.0.

### To test:
Review the changes. To make it less confusing, just look at the changes
after the reverts:

https://github.com/mantidproject/mantid/pull/40077/commits/df7b4d10bbbdcb4139aa07763059dfe8f601e3b5
This commit simply removes all osx-64 stages.