Change some restrictions on input validation of SANS Polarization Corrections Algorithm (#40100)
### Description of work
This PR fixes #40099 . More details can be found in issue. 

Additionally, as this algorithm is being tested, I have tried to
increase the flexibility for diagnostics in this algorithm, adding a
check for avoiding deleting partial workspaces if the algorithm fails,
as well as allowing to set initial values for some of the fitting
properties.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes #40099. <!-- One line per closed issue. -->
- Test are passing. 
- Contact me for test data.

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

Co-authored-by: Jonathan Haigh <35813666+jhaigh0@users.noreply.github.com>