Fix all the broken help links (#40104)
### Description of work
A lot of the interfaces and settings etc had broken help links that were
missed when we moved to the new help window. This fixes them. Here is a
list of what was broken:

- plot config dialog help
- release notes
- DrillExportDialog (drill export settings window)
- Main mantid settings
- sliceviewer
- powder diffraction reduction -> advanced setup -> the help buttons at
the bottom - there are two.
- script repository
- instrument view
- ISIS SANS

Closes #40025 

### To test:
Linux packages are here:
`mamba install -c thomashampson/label/help-link-fix`
Check that all the above examples work.


<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->