Fix bug in Event Filter GUI (#40086)
### Description of work
This PR fix a few bugs with line edits found in #40026 and #40000. 
I have added two improvements to fix the bugs and reduce the chance of
bugs in the future with this interface:
- Change the four sliders to two `RangeMarkers` (same as those we use in
fitting and inelastic interfaces). This make the usage less clunky, the
code substantially simpler and reduced and hopefully more stable.
- Add a custom double validator for the line edits, overriding the
`fixup` method to provide a corrected value when the input is not valid.
This prevents the input of wrong values or the edit keeping wrong values
when it goes out of focus (which was the cause for the bug).
- Additionally, manual testing instructions and interface screenshots
are updated.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes #40026. <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
- Build docs, make sure all new images are rendered properly.
- Build developer docs, and follow new manual testing instructions for
the interface. Everything should work.


<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

Co-authored-by: James Clarke <139879523+jclarkeSTFC@users.noreply.github.com>