Fix colorbar axis edit (#40179)
### Description of work

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

This PR fixes an issue that occurs when trying to change the limits of a
colorbar on a contour plot.

I was also intending to fix the error that occurs when pressing `c` with
a colorbar scale set to log (#40136), but the fix required (other than
disabling) is too complex to put in just before release.

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

1) Plot any data
2) 3D plot, contour
3) Double click on colorbar axis, change limits.