Stop messages from crashing mantid on macOS ARM (#40180)
### Description of work

`#if`s out the check to see if the OS is supported (as it seems to be
lying) on Apple ARM machines since it causes a crash whenever anything
would try to post messages to the SystemTray.

I'm reasonably sure these have never worked on macOS anyway, so just
stopping them from running isn't a great loss.
I've made attempts to see if there was a quick fix to actually make them
work, but didn't find one - and any changes would be too big to take
place this close to release anyway.

Closes #40176 
Closes #40177 


### To test:
1. Open Workbench
2. Attempt to Open the `ILL -> DrILL` interface (with the facility set
to ISIS).
3. Throw a syntax error from the script editor.
4. Attempt to open the `Direct -> DNS Reduction` interface.

*This does not require release notes* because nothing should change from
the user's perspective.