Peak overlay in Instrument View (#40101)
Peak overlay for new Instrument View. Peaks workspaces with the same
instrument as the main workspace will show on the left of the interface,
and can overlaid by selecting their checkbox. If a peaks workspace is
overlaid and a detector is selected in the projection with a peak, then
that peak will also be overlaid on the 1D plot with the HKL value.

Closes #40043  <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

To aid testing, you can have both the old and new instrument view
interfaces open at the same time, and pick/delete peaks in the old one.
The new one should update.

You can't pick peaks in the new interface yet.

<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->