Use columnArray for setting up New Instrument View model (#40193)
### Description of work

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->
Use improved methods `ColumnArray` and argument `PickOneDetectorID` in
`CreateDetectorTable` to simplify the creating of the model.
The performance improvement comes mainly from the creating of the numpy
arrays happening on C++ side in `ColumnArray()`.
The speedup I recorded on my machine is less than what I was expecting:
I get around 5X speedup.

Closes #39867 . <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
Open instrument view and check that is unchanged and working as
expected.
If you want, use pythons `time.time()` to measure the before and after
speedup of the setup.