Remove peaks from workspace directly
By calling an algorithm we rely on the workspace getting updated, then
the ADS callback coming back into the peak overlay and updating the
workspace. If deleting lots of peaks in a short space of time it's luck
whether or not the workspace has been updated. If out of date then the
overlay will try and get an invalid index, interface will disappear.