Refactor Polarization Corrections Tests (#39961)
### Description of work
This is a maintenance issue. Many of the tests for the sans polarization
corrections algorithms can share similar test data, however every tests
for each polarization algorithm has its own set of helper functions to
create a workspace group suitable for polarization algorithms. I have
tried to unify all this helper functions into a new file
`PolarizationCorrectionsTestsUtils`, which is already recycled from a
helper file used in the analyser efficiency tests.
The two main changes are the inclusion of a helper function
`createPolarizedTestGroup` that can be used in all the polarization
correction tests and a struct to hold the main parameters of the
workspaces and efficiency functions.

Note: I have set the release to 6.15 as there is no hurry to get this
reviewed before code freeze for 6.14 if there are more important PRs to
review.

No associated issue. 

### To test:
- All tests are passing. 

__Release notes no necessary as it is internal change for tests only__