Refactor CreateDetectorTable (#40194)
### Description of work
Refactored `CreateDetectorTable` algorithm to make it a bit cleaner and
easier to modify in the future. All I did was making the functions part
of the class so that the arguments don't need to be passed on from
function to function.

### To test:
Try CreateDetectorTable with a peaks workspace and a normal matrix
workspace and check that the table looks fine.

Co-authored-by: Andy Bridger <andy.bridger@stfc.ac.uk>