Measure angle between two picked detectors
The angle is between the Q lab vectors for each detector. The current
interface only gives you the Q lab vector for a peak, but you can
work out the direction from the detector position. All peaks on a
detector will have the same Q lab direction but with different
magnitude.