Allow Linux to build with clang (#40225)
While we don't support building with clang on Linux there is no reason
it shouldn't work. This did compile previously. So this fixes a couple
of missing headers and a template specialization for `PropertyWithValue`
after which you should be able to use clang on Linux to compile and run
mantid workbench again.

The unit tests do compile and run with clang but I found about 20 fail.
I haven't looked too closely as to why.

To build with clang on linux first install clang in your conda
environment with whatever tool you are using, _e.g._

```
micromamba install clangxx==18.1.8
```

I installed version 18 as that is the version OSX
[uses](https://github.com/mantidproject/mantid/blob/c1052ef805e0c48a4878f22cc3c07f0bd48501d9/conda/recipes/conda_build_config.yaml#L6).

Then run cmake specifying which compilers to use with a clean build
directory

```
CC=clang CXX=clang++ cmake ...
```

and check that cmake found the expected compiler.

You should then be able to compile and run mantid workbench as normal.

My thought is this may help with fixing some issues in #39159 where we
could get it working on Linux first.

### To test:

Build and start mantid workbench as shown above.

*This does not require release notes* because it's for developers only.