Update the environment log urls and names since rattler build for the dependency spotter script  (#40241)
### Description of work

Updated the name of the base environment log and the url to the logs in
the dependency spotter script.
Also had to update the regex logic which was only retrieving the first
environment it found instead of all of them. I think what happened is
that originally the environment files appeared in separate lines of
html, now they come all in one, so I switched to using `re.findall`

Also, the windows logs seem to be in the wrong place
https://github.com/mantidproject/mantid/issues/40214#issuecomment-3457316170
This pr assumes that this will be fixed.

Closes #40214

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

Just play around with the script using recent nightly runs. There is no
backwards compatibility for before the rattler-build changes