Correct Abins Lagrange resolution model (#40229)
Abins applies instrumental broadening to its simulated spectra. For the
Lagrange instrument and Cu(220) monochromator, this is modelled as a
Gaussian with width depending on a polynomial and a flat low-energy
limit.

A unit conversion was missing between the frequencies (in wavenumber)
and lower-bound of polynomial fit (in meV), causing the low-energy limit
to be engaged at the wrong energy.

### Description of work

The discrepancy between the model as fitted/described and implementated
was discovered by Rastislav Turanyi while working on ResINS. The origin
is a missing unit conversion.

As discovered in the thread below, the "corrected" function still has a
concerning discontinuity in it; to get a simple and reasonable
behaviour, we instead implement a basic floor function.

The LAGRANGE models could use a closer look in general, but for now this
is clearly an improvement.

### To test:

A new unit test is included