Delete FindPoco.cmake (#40016)
Poco comes with its own cmake module, we can use that instead of our own
home brew. The only thing is that it defines `UNICODE` and `_UNICODE` as
compiler arguments, which causes chaos on Windows, since we would then
have to use `wstring` in several places. Hence in the `cmake` the
`UNICODE` and `_UNICODE` flags are removed.

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

Check CI and package build, try opening Workbench and running some
algorithms.