Make use of a new mantid-test Anaconda channel (#40256)
A new mantid-test channel has been created, which can be published to
using the nightly deployment Jenkinsfile. It is intended to be used for
two purposes:
1. For developers to publish to when using the
[build_branch](https://builds.mantidproject.org/job/build_branch/)
pipeline. This will avoid clogging up the `mantid` channel with test
packages and help us stay under the allowed storage limit.
2. For uploading the smoke testing packages for release testing.
Both the above use cases will also avoid the need to manually delete
packages from the main `mantid` channel, which is too prone to user
error (e.g. one can easily delete production packages by mistake).
Developer documentation has been updated accordingly.

Below is a summary of how the new mechanism works.
- The `ANACONDA_CHANNEL` Jenkins pipeline parameter is now a choice
parameter. It used to be a string parameter. A choice parameter makes
more sense since we only store a limited number of credentials in
Jenkins, so anything outside of the available choices will not work.
- The `mantid` channel is only available as a choice if the branch name
is `main` or `release-next`.
- The `mantid-test` channel is available for all branch names.
- The `mantid-ornl` channel is available for only `ornl` branches.
- Channel names and credential IDs are mapped within the jenkinsfile, so
the correct credentials are used for each channel.

Closes #40017

### To test:
The following Jenkins job was set to publish to the new mantid-test
channel (look a the parameter values set in the job):
https://builds.mantidproject.org/job/Core_Team_test_pipeline/416/
You can verify that it did upload to the correct label by visiting the
channel:
https://anaconda.org/mantid-test/repo/files?label=tom-test-one